+++ js/objects/background.js
@@ -15,6 +15,7 @@
         size: ?, (string, css background-size for image resizing, shrinking, repeating)
         repeat: no-repeat|repeat-x|repeat-y|repeat,
     },
+    svg_filters: ?, (string, list of <filter> tag)
 }*/
 
 function Background(glob_data) {
@@ -36,6 +37,7 @@ function Background(glob_data) {
             SIZE: "",
             REPEAT: "no-repeat",
         },
+        SVG_FILTERS: "",
     };
 
 
@@ -121,6 +123,13 @@ function Background(glob_data) {
 
         }
 
+        //svg filter
+        if ( imports.utils.IsUndefined(data.svg_filters) && !(ignore_undefined === "IGNORE_UNDEFINED") ) {data.svg_filters = this.DEFAULTS.SVG_FILTERS;}
+        if ( !imports.utils.IsUndefined(data.svg_filters) && data.svg_filters !== "" && (!imports.utils.IsAString(data.svg_filters) || data.svg_filters.includes("<script>") || !data.svg_filters.includes("<filter") || !data.svg_filters.includes("</filter>")) ) {
+            CustomLog("warn",`Visualizer object: Invalid svg filters! Set to "${this.DEFAULTS.SVG_FILTERS}".`); //do not detect html errors!
+            data.svg_filters = this.DEFAULTS.SVG_FILTERS;
+        }
+
         return data;
 
     }
@@ -189,6 +198,25 @@ function Background(glob_data) {
                     break;
             }
             this.element.style.backgroundSize = this.data.background.size;//size
+            //svg filters
+            if (this.data.svg_filters !== "") {
+                this.svg_filter_div.children[0].children[0].innerHTML = "";
+                let str = "";
+                let filters = this.data.svg_filters.split("[#]");
+                for (let i = 0; i < filters.length; i++) {
+                    let id = `${this.data.id}-FILTER-${i}`;
+                    //load filter
+                    filters[i] = filters[i].replace(/id=".*?"/,"").replace("<filter",`<filter id="${id}"`);
+                    this.svg_filter_div.children[0].children[0].innerHTML += filters[i];
+                    //add in css
+                    let strToAdd = `url('#${id}')`;
+                    str += strToAdd;
+                    //spacing
+                    if (i < filters.length-1) str += " ";
+                }
+                this.element.style.filter = str;
+                console.log(str);
+            }
 
             //SET BACKGROUND TO SCREEN SIZE
             this.element.style.width = screen.width + "px";
@@ -216,6 +244,12 @@ function Background(glob_data) {
     this.element.style.left = 0;
     this.element.style.display = "inline-block";
     this.element.style.overflow = "hidden";
+
+    //svg filters
+    this.svg_filter_div = document.createElement("div");
+    document.body.appendChild(this.svg_filter_div);
+    this.svg_filter_div.innerHTML = "<svg xmlns='http://www.w3.org/2000/svg' version='1.1'><defs></defs></svg>";
+
 
 
 
@@ -248,6 +282,7 @@ function Background(glob_data) {
 
         this.parameters = {
             layer: null,
+            svg_filters: null,
         };
 
         //layer
@@ -366,6 +401,21 @@ function Background(glob_data) {
                 this_object.updateData(updated_data);
             }
         );
+        
+        //svg filter
+        this.parameters.svg_filters = new imports.ui_components.UIParameterString(
+            this.parameter_container,
+            "SVG Filters (advanced, read help)",
+            this.data.svg_filters,
+            () => {
+                this.updateData({
+                    id: this.data.id,
+                    svg_filters: this.parameters.svg_filters.value,
+                });
+            }
+        );
+        this.parameters.svg_filters.help_string = help.parameter.object.general.svg_filters;
+
     }