+++ js/objects/image.js
@@ -22,6 +22,7 @@
     },
     border_radius: ?, (string, css border-radius)
     box_shadow: ?, (string, css box-shadow)
+    svg_filters: ?, (string, list of <filter> tag)
 }*/
 
 function Image(glob_data) {
@@ -50,6 +51,7 @@ function Image(glob_data) {
         },
         BORDER_RADIUS: "",
         BOX_SHADOW: "",
+        SVG_FILTERS: "",
     };
 
 
@@ -185,6 +187,13 @@ function Image(glob_data) {
             data.box_shadow = this.DEFAULTS.BOX_SHADOW;
         }
 
+        //svg filter
+        if ( imports.utils.IsUndefined(data.svg_filters) && !(ignore_undefined === "IGNORE_UNDEFINED") ) {data.svg_filters = this.DEFAULTS.SVG_FILTERS;}
+        if ( !imports.utils.IsUndefined(data.svg_filters) && data.svg_filters !== "" && (!imports.utils.IsAString(data.svg_filters) || data.svg_filters.includes("<script>") || !data.svg_filters.includes("<filter") || !data.svg_filters.includes("</filter>")) ) {
+            CustomLog("warn",`Visualizer object: Invalid svg filters! Set to "${this.DEFAULTS.SVG_FILTERS}".`); //do not detect html errors!
+            data.svg_filters = this.DEFAULTS.SVG_FILTERS;
+        }
+
         return data;
 
     }
@@ -259,6 +268,25 @@ function Image(glob_data) {
             this.element.style.backgroundSize = this.data.background.size;//background size
             this.element.style.borderRadius = this.data.border_radius;//
             this.element.style.boxShadow = this.data.box_shadow;
+            //svg filters
+            if (this.data.svg_filters !== "") {
+                this.svg_filter_div.children[0].children[0].innerHTML = "";
+                let str = "";
+                let filters = this.data.svg_filters.split("[#]");
+                for (let i = 0; i < filters.length; i++) {
+                    let id = `${this.data.id}-FILTER-${i}`;
+                    //load filter
+                    filters[i] = filters[i].replace(/id=".*?"/,"").replace("<filter",`<filter id="${id}"`);
+                    this.svg_filter_div.children[0].children[0].innerHTML += filters[i];
+                    //add in css
+                    let strToAdd = `url('#${id}')`;
+                    str += strToAdd;
+                    //spacing
+                    if (i < filters.length-1) str += " ";
+                }
+                this.element.style.filter = str;
+                console.log(str);
+            }
         }
 
 
@@ -280,6 +308,11 @@ function Image(glob_data) {
     this.element.style.position = "absolute";
     this.element.style.display = "inline-block";
     this.element.style.overflow = "hidden";
+
+    //svg filters
+    this.svg_filter_div = document.createElement("div");
+    document.body.appendChild(this.svg_filter_div);
+    this.svg_filter_div.innerHTML = "<svg xmlns='http://www.w3.org/2000/svg' version='1.1'><defs></defs></svg>";
 
 
 
@@ -318,6 +351,7 @@ function Image(glob_data) {
             rotation: null,
             border_radius: null,
             box_shadow: null,
+            svg_filters: null,
         };
 
         //layer
@@ -613,6 +647,20 @@ function Image(glob_data) {
             }
         );
         this.parameters.box_shadow.help_string = help.parameter.object.general.shadow;
+
+        //svg filter
+        this.parameters.svg_filters = new imports.ui_components.UIParameterString(
+            this.parameter_container,
+            "SVG Filters (advanced, read help)",
+            this.data.svg_filters,
+            () => {
+                this.updateData({
+                    id: this.data.id,
+                    svg_filters: this.parameters.svg_filters.value,
+                });
+            }
+        );
+        this.parameters.svg_filters.help_string = help.parameter.object.general.svg_filters;
 
     }