+++ js/objects/particle_flow.js
@@ -21,6 +21,7 @@
     spawn_probability: ?, (float) //probability to spawn a particle at each test (0: none, 1: full)
     spawn_tests: ?, (int >=1) //how many spawn tests are done at every frame
     color: ?, (string: hex, rgb, rgba)
+    svg_filters: ?, (string, list of <filter> tag)
 }*/
 
 function ParticleFlow(glob_data) {
@@ -49,6 +50,7 @@ function ParticleFlow(glob_data) {
         SPAWN_PROBABILITY: 0.75,
         SPAWN_TESTS: 1,
         COLOR: "#ffffff",
+        SVG_FILTERS: "",
     }
 
 
@@ -163,6 +165,13 @@ function ParticleFlow(glob_data) {
             data.color = this.DEFAULTS.COLOR;
         }
 
+        //svg filter
+        if ( imports.utils.IsUndefined(data.svg_filters) && !(ignore_undefined === "IGNORE_UNDEFINED") ) {data.svg_filters = this.DEFAULTS.SVG_FILTERS;}
+        if ( !imports.utils.IsUndefined(data.svg_filters) && data.svg_filters !== "" && (!imports.utils.IsAString(data.svg_filters) || data.svg_filters.includes("<script>") || !data.svg_filters.includes("<filter") || !data.svg_filters.includes("</filter>")) ) {
+            CustomLog("warn",`Visualizer object: Invalid svg filters! Set to "${this.DEFAULTS.SVG_FILTERS}".`); //do not detect html errors!
+            data.svg_filters = this.DEFAULTS.SVG_FILTERS;
+        }
+
         return data;
 
     }
@@ -211,6 +220,25 @@ function ParticleFlow(glob_data) {
             this.element.width = this.data.width;//width
             this.element.height = this.data.height;//height
             // ##############
+            //svg filters
+            if (this.data.svg_filters !== "") {
+                this.svg_filter_div.children[0].children[0].innerHTML = "";
+                let str = "";
+                let filters = this.data.svg_filters.split("[#]");
+                for (let i = 0; i < filters.length; i++) {
+                    let id = `${this.data.id}-FILTER-${i}`;
+                    //load filter
+                    filters[i] = filters[i].replace(/id=".*?"/,"").replace("<filter",`<filter id="${id}"`);
+                    this.svg_filter_div.children[0].children[0].innerHTML += filters[i];
+                    //add in css
+                    let strToAdd = `url('#${id}')`;
+                    str += strToAdd;
+                    //spacing
+                    if (i < filters.length-1) str += " ";
+                }
+                this.element.style.filter = str;
+                console.log(str);
+            }
         }
 
 
@@ -234,6 +262,10 @@ function ParticleFlow(glob_data) {
     this.element.style.display = "inline-block";
     this.element.style.overflow = "hidden";
 
+    //svg filters
+    this.svg_filter_div = document.createElement("div");
+    document.body.appendChild(this.svg_filter_div);
+    this.svg_filter_div.innerHTML = "<svg xmlns='http://www.w3.org/2000/svg' version='1.1'><defs></defs></svg>";
 
 
 
@@ -279,6 +311,7 @@ function ParticleFlow(glob_data) {
             spawn_probability: null,
             spawn_tests: null,
             color: null,
+            svg_filters: null,
         };
 
         //layer
@@ -595,6 +628,21 @@ function ParticleFlow(glob_data) {
             }
         );
         this.parameters.color.help_string = help.parameter.object.particles.ptcl_color;
+        
+        //svg filter
+        this.parameters.svg_filters = new imports.ui_components.UIParameterString(
+            this.parameter_container,
+            "SVG Filters (advanced, read help)",
+            this.data.svg_filters,
+            () => {
+                this.updateData({
+                    id: this.data.id,
+                    svg_filters: this.parameters.svg_filters.value,
+                });
+            }
+        );
+        this.parameters.svg_filters.help_string = help.parameter.object.general.svg_filters;
+
     }